code/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0
pyyaml>=6.0
